Correct the XML documentation for SystemTextJson implementation of NodaConverterBase.CanConvert
It looks like System.Text.Json doesn't need or expect a converter to advertise that it can handle nullable value types. (The default implementations don't do so.)

It's entirely possible that we should actually just not override the method at all here, and only do so in NodaDateTimeZoneConverter, but I don't want to change the implementation right now.